refactor(martech): move consent detection into separate file
